Remove unused error message
I considered leaving this error message because being so granular might
be confusing. This takes away the granularity to direct you to
the guard result. Once you know it's the guard result you can likely
figure out the problem from there.

I ultimately decided to stick with the granular errors and if devs get
confused I think adding stacktrace information to the response is the
real solution.