refactor: narrow scope of unsafe blocks
Unsafe code is scary so we should use the tag to only mark the things
that are truly unsafe, not just large blocks of code.

Mutably accessing the ref cells is the dangerous thing so I wrapped only
that part with the unsafe block.

This makes it more clear which parts of the code are still safe.